requirements.txt
beautifulsoup4
requests
redis
PyYAML